Improved circle adding
Recoded circles so that no longer a 64x64 byte block is used, but an
offset, a cordlength and actual circle data. Taking this offset and
lengths into consideration speeds up to code drastically.

In worst case scearios ( all 4 circles in the middel of the screen)  the
new code takes only 75% of the time of the old code. Which corresponds
to the size difference between the square and circle

In best case scenario (all 4 circles in the corner) the speed up is
still 82% of the timeof the old code